requirements.txt
python-dotenv
requests
Pillow
pydub
pyttsx3
edge-tts
PySide6
fastapi
uvicorn[standard]
